desktop_app/requirements.txt
PyQt6==6.7.1
pyserial==3.5